Moved login logic to makeRequest function
