+++ sdk/index.js
@@ -1,5 +1,5 @@
 // istanbul ignore file
-import { initialize, login } from './interfaces';
+import { initialize, login, getToken } from './interfaces';
 import { getParameters } from './configuration';
 
-export { initialize, login, getParameters };
+export { initialize, login, getToken, getParameters };
