+++ sdk/requests/__tests__/index.test.js
@@ -4,47 +4,7 @@ import { getParameters } from '../../configuration';
 
 jest.mock('../../configuration');
 
-const mockLinkingOpenUrl = jest.fn(() => Promise.resolve());
-
-jest.mock('react-native/Libraries/Linking/Linking', () => ({
-  openURL: mockLinkingOpenUrl,
-}));
-
 afterEach(() => jest.clearAllMocks());
-
-describe('login', () => {
-  it('calls login with clientId and redirectUri', async () => {
-    getParameters.mockReturnValue({
-      clientId: 'clientId',
-      redirectUri: 'redirectUri',
-    });
-    await makeRequest(REQUEST_TYPES.LOGIN);
-    expect(mockLinkingOpenUrl).toHaveBeenCalledWith(
-      'https://auth-testing.iduruguay.gub.uy/oidc/v1/authorize?scope=openid&response_type=code&client_id=clientId&redirect_uri=redirectUri',
-    );
-  });
-
-  it('calls login without clientId and with redirectUri', async () => {
-    getParameters.mockReturnValue({
-      clientId: '',
-      redirectUri: 'redirectUri',
-    });
-    const response = await makeRequest(REQUEST_TYPES.LOGIN);
-    expect(response).toBe('');
-    expect(mockLinkingOpenUrl).not.toHaveBeenCalled();
-  });
-
-  it('calls login with clientId and without redirectUri', async () => {
-    getParameters.mockReturnValue({
-      clientId: 'clientId',
-      redirectUri: '',
-    });
-    await makeRequest(REQUEST_TYPES.LOGIN);
-    expect(mockLinkingOpenUrl).toHaveBeenCalledWith(
-      'https://auth-testing.iduruguay.gub.uy/oidc/v1/authorize?scope=openid&response_type=code&client_id=clientId&redirect_uri=',
-    );
-  });
-});
 
 jest.mock('react-native-ssl-pinning', () => ({
   fetch: jest.fn(),
@@ -141,6 +101,5 @@ describe('default', () => {
   it('calls default', async () => {
     const response = await makeRequest('default');
     expect(response).toBe('default value');
-    expect(mockLinkingOpenUrl).not.toHaveBeenCalled();
   });
 });
