+++ sdk/requests/endpoints.js
@@ -2,8 +2,8 @@ import { getParameters } from '../configuration';
 
 const tokenEndpoint = 'https://auth-testing.iduruguay.gub.uy/oidc/v1/token';
 
-const loginEndpoint = clientId => {
-  const { redirectUri } = getParameters();
+const loginEndpoint = () => {
+  const { redirectUri, clientId } = getParameters();
   return `https://auth-testing.iduruguay.gub.uy/oidc/v1/authorize?scope=openid&response_type=code&client_id=${clientId}&redirect_uri=${redirectUri}`;
 };
